+++ tests/test_wsknn.py
@@ -38,10 +38,25 @@ def test_wsknn_invalid_type_exception():
         model.fit(sessions=sessions, items=items)
 
 
+def test_wsknn_invalid_datatypes_exception():
+    sessions = {
+        'A': [[1, 2, 3], [10, 20, 30]]
+    }
+
+    items = {
+        '1': [['A'], [10]]
+    }
+
+    model = WSKNN()
+
+    with pytest.raises(TypeError):
+        model.fit(sessions=sessions, items=items)
+
+
 def test_wsknn_timestamp_type_exception():
     sessions = {
         'A': [[1, 2, 3],
-              [10, 20, 30]]
+              ['10', '20', '30']]
     }
 
     items = {
@@ -81,7 +96,8 @@ def test_wsknn_flow1():
 
     some_sessions = [[[1], [100]], [[2, 3], [200, 300]]]
 
-    expected_recommendations = [[(2, 2.0), (3, 2.0), (4, 2.0), (5, 2.0)], [(4, 2.5), (5, 2.5), (1, 1.25)]]
+    expected_recommendations = [[(2, 2.0), (3, 2.0), (4, 2.0), (5, 2.0)],
+                                [(4, 2.5), (5, 2.5), (1, 1.25)]]
 
     model = WSKNN(return_events_from_session=False)
     model.fit(sessions, items)
@@ -181,7 +197,8 @@ def test_wsknn_no_items():
 
     some_sessions = [[[1], [100]], [[2, 3], [200, 300]]]
 
-    expected_recommendations = [[(2, 2.0), (3, 2.0), (4, 2.0), (5, 2.0)], [(4, 2.5), (5, 2.5), (1, 1.25)]]
+    expected_recommendations = [[(2, 2.0), (3, 2.0), (4, 2.0), (5, 2.0)],
+                                [(4, 2.5), (5, 2.5), (1, 1.25)]]
 
     model = WSKNN(return_events_from_session=False)
     model.fit(sessions)
@@ -189,3 +206,61 @@ def test_wsknn_no_items():
     for idx, sess in enumerate(some_sessions):
         recomms = model.recommend(sess)
         assert recomms == expected_recommendations[idx]
+
+
+def test_batch_recommendations():
+    sessions = {
+        'a': [
+            [1, 2, 3, 4, 5],
+            [1, 2, 3, 4, 5]
+        ],
+        'b': [
+            [2, 3, 4, 5],
+            [10, 11, 12, 13]
+        ]
+    }
+
+    some_sessions = {
+        'x': [[1], [100]],
+        'y': [[2, 3], [200, 300]]
+    }
+
+    expected_recommendations = {
+        'x': [(2, 2.0), (3, 2.0), (4, 2.0), (5, 2.0)],
+        'y': [(4, 2.5), (5, 2.5), (1, 1.25)]
+    }
+
+    model = WSKNN(return_events_from_session=False)
+    model.fit(sessions)
+
+    recs = model.recommend(some_sessions)
+    for _k, _v in recs.items():
+        assert expected_recommendations[_k] == _v
+
+
+def test_wsknn_random_sampling():
+    # Type of sessions is str, type of items is int
+    sessions = {
+        'a': [
+            [1, 2, 3, 4, 5],
+            [1, 2, 3, 4, 5]
+        ],
+        'b': [
+            [2, 3, 4, 5],
+            [10, 11, 12, 13]
+        ]
+    }
+
+    items = {
+        1: [['a'], [1]],
+        2: [['a', 'b'], [2]],
+        3: [['a', 'b'], [3]],
+        4: [['a', 'b'], [4]],
+        5: [['a', 'b'], [5]]
+    }
+
+    model = WSKNN(return_events_from_session=False,
+                  sampling_strategy='random')
+    model.fit(sessions, items)
+
+    assert isinstance(model, WSKNN)